% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-s-score.R
\name{calculateSScore}
\alias{calculateSScore}
\title{Calculate S-Score for Mean Reverting Stocks}
\usage{
calculateSScore(returns, nr_pc, h, l, b_sensitivity)
}
\arguments{
\item{returns}{A numeric matrix, the returns data.}

\item{nr_pc}{An integer, the number of eigen portfolios to extract.}

\item{h}{An integer, the number of recent historical days to be used.}

\item{l}{An integer, the number of preceding days used for the model.}

\item{b_sensitivity}{A numeric value, sensitivity for checking if model b coefficient is less than 1 minus this value.}
}
\value{
A list containing:
* `s`: The computed S-Score for each stock.
* `is_mean_reverting`: A logical vector indicating whether each stock is mean reverting.
}
\description{
This function calculates the S-Score for each stock which is proven to be mean reverting.
It first decomposes the returns and estimate key coefficients of each model. Based on 
these coefficients and mean reverting condition an S-score is calculated.
}
\examples{
#Example data
ret <- matrix(rnorm(25), 5, 5)
num_pc <- 2
historical <- 4
preceding <- 1
sensitivity <- 0.1
#Use the function
calculateSScore(ret, num_pc, historical, preceding, sensitivity)
}
