% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combining-forecasts.R
\name{combinePrediction}
\alias{combinePrediction}
\title{Combine Predictions}
\usage{
combinePrediction(strong_predictions, weak_predictions, alpha)
}
\arguments{
\item{strong_predictions}{A numeric matrix, the strong predictions.}

\item{weak_predictions}{A numeric matrix, the weak predictions.}

\item{alpha}{Numeric value, the quantile threshold used in the 
`combineDailyPredictions` function.}
}
\value{
A numeric matrix, the combined predictions. The column names of the
returned matrix will be the same as the column names of the strong predictions.
}
\description{
This function combines strong and weak stock predictions. It applies the 
`combineDailyPredictions` function for each column of the data.
}
\examples{
#Example data
strong_predictions <- matrix(runif(25), 5, 5)
weak_predictions <- matrix(runif(25), 5, 5)
alpha <- 0.8
#Use the function
combinePrediction(strong_predictions, w_pred, alpha)
}
