% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross-sectional-regression-clustering.R
\name{constructClusters}
\alias{constructClusters}
\title{Construct Clusters Based on Hierarchical Clustering}
\usage{
constructClusters(rho, k, min_size)
}
\arguments{
\item{rho}{A numeric matrix, the similarity matrix. It is often a correlation matrix.}

\item{k}{An integer, the desired number of clusters.}

\item{min_size}{An integer, the minimum desired size of the clusters.}
}
\value{
A list where each element is a vector of indices representing a cluster.
}
\description{
This function applies hierarchical clustering to construct k clusters based on a similarity matrix `rho` (e.g. correlation matrix). 
If the minimum size of any cluster is less than a specified `min_size`, the function will reduce the number of clusters (`k`) until this condition is met.
}
\examples{
#Example data
rho <- matrix(rnorm(16), 4, 4)
k <- 2
min_size <- 2
#Use the function
constructClusters(rho, k, min_size)

}
