% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-eigen.R
\name{constructEigenPortfolios}
\alias{constructEigenPortfolios}
\title{Construct Eigen Portfolios}
\usage{
constructEigenPortfolios(returns)
}
\arguments{
\item{returns}{A numeric matrix, the returns data.}
}
\value{
A list containing:
* `rho`: The correlation matrix for the returns.
* `portfolios`: The eigen portfolios.
* `returns`: The returns of each eigen portfolio.
* `values`: The eigenvalues of the correlation matrix.
}
\description{
This function constructs eigen portfolios from a matrix of returns. It first constructs 
a correlation matrix (rho) of the returns. Then it computes the eigenvalues and eigenvectors 
of the correlation matrix. The eigenvectors are then normalized by the standard deviation 
of the returns for each stock. The function also computes the returns for each eigen portfolio.
}
\examples{
#Example data
ret <- matrix(rnorm(25), 5, 5)
#Use the function
constructEigenPortfolios(ret)
}
