% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-eigen.R
\name{constructRho}
\alias{constructRho}
\title{Construct Correlation Matrix}
\usage{
constructRho(returns)
}
\arguments{
\item{returns}{A numeric matrix, the returns data.}
}
\value{
A correlation matrix with row-wise scaled data.
}
\description{
This function constructs a correlation matrix (rho) from a matrix of returns.
Each row of the matrix is scaled to have a mean of 0 and standard deviation of 1 
before calculating the correlation matrix.
}
\examples{
#Example data
returns <- matrix(rnorm(25), 5, 5)
#Use the function
constructRho(ret)
}
