% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross-sectional-regression-vw.R
\name{constructWeightedReturns}
\alias{constructWeightedReturns}
\title{Construct Weighted Returns}
\usage{
constructWeightedReturns(returns, volume, d, divide)
}
\arguments{
\item{returns}{A numeric matrix, the returns data.}

\item{volume}{A numeric matrix, the volume data.}

\item{d}{An integer, the width for calculating the rolling mean of the volume.}

\item{divide}{A logical value, if TRUE, the function divides the product of returns and rolling mean volume by the volume. If FALSE, it divides the product of returns and volume by the rolling mean volume.}
}
\value{
A matrix of the same dimensions as returns and volume, where each entry corresponds to the weighted return of a stock on a particular day.
}
\description{
This function generates weighted returns based on the ratio of rolling mean volume to actual volume, 
or vice-versa depending on the value of the divide parameter.
}
\examples{
#Example data
returns <- matrix(rnorm(25), 5, 5)
volume <- matrix(rnorm(25), 5, 5)
d <- 2
divide <- TRUE
#Use the function
constructWeightedReturns(returns, volume, d, divide)
}
