% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross-sectional-regression.R
\name{crossSectionalRegression}
\alias{crossSectionalRegression}
\title{Cross-Sectional Regression}
\usage{
crossSectionalRegression(returns, start, end, h, nr_pc)
}
\arguments{
\item{returns}{A numeric matrix, the returns data.}

\item{start}{An integer, the start of the days range for which to perform regression.}

\item{end}{An integer, the end of the days range for which to perform regression.}

\item{h}{An integer, the number of recent historical days to be used.}

\item{nr_pc}{An integer, the number of eigen portfolios to extract.}
}
\value{
A numeric matrix of predicted returns where each row corresponds to a day in the given `start:end` range and each column corresponds to a stock.
}
\description{
This function performs cross-sectional regression on returns on a range of days,
using recent historical data. An eigenportfolio is first extracted from the historical data.
Then, a linear model is fit between the returns and the eigenportfolio for each day in the range.
The function returns the negative residuals from these models as predicted returns. 
This function utilizes parallel computing for efficiency.
}
\examples{
#Example data
returns <- matrix(rnorm(25), 5, 5)
start <- 2
end <- 5
h <- 4
nr_pc <- 2
#Use the function
crossSectionalRegression(returns, start, end, h, nr_pc)
}
