% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross-sectional-regression-kpca.R
\name{crossSectionalRegressionKPCA}
\alias{crossSectionalRegressionKPCA}
\title{Cross-Sectional Regression with Kernel Principal Component Analysis (KPCA)}
\usage{
crossSectionalRegressionKPCA(returns, start, end, h, nr_pc, kernel, kpar)
}
\arguments{
\item{returns}{A numeric matrix, the returns data.}

\item{start}{An integer, the start of the days range for which to perform regression.}

\item{end}{An integer, the end of the days range for which to perform regression.}

\item{h}{An integer, the number of recent historical days to be used.}

\item{nr_pc}{An integer, the number of principal components to use in KPCA.}

\item{kernel}{A character string, the kernel to be used in KPCA.}

\item{kpar}{A list, the list of hyperparameters for the kernel.}
}
\value{
A numeric matrix of predicted returns where each row corresponds to a day in the given `start:end` range and each column corresponds to a stock.
}
\description{
This function performs cross-sectional regression on a range of days, using recent historical data. 
The analysis uses Kernel Principal Component Analysis (KPCA) with the provided kernel and its parameters. 
The parallelized function applies Kernel PCA to the scaled returns to extract principal components, then fits a 
linear model and calculates residuals as predicted returns.
}
\examples{
#Example data
returns <- matrix(rnorm(20), 4, 5)
start <- 1
end <- 4
h <- 2
nr_pc <- 2
kernel <- "rbfdot"
kpar <- list(sigma = 0.1)
#Use the function
crossSectionalRegressionKPCA(returns, start, end, h, nr_pc, kernel, kpar)

}
