% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross-sectional-regression-overtrade.R
\name{crossSectionalRegressionOvertraded}
\alias{crossSectionalRegressionOvertraded}
\title{Cross-Sectional Regression for Overtraded Volume}
\usage{
crossSectionalRegressionOvertraded(
  start,
  end,
  volume,
  returns,
  h,
  nr_pc_v,
  alpha,
  nr_pc
)
}
\arguments{
\item{start}{An integer, the start of the day range for which to perform regression.}

\item{end}{An integer, the end of the day range for which to perform regression.}

\item{volume}{A numeric matrix, the volume data.}

\item{returns}{A numeric matrix, the returns data.}

\item{h}{An integer, the number of recent historical days to be used.}

\item{nr_pc_v}{An integer, the number of eigen portfolios to extract from the volume data.}

\item{alpha}{A numeric, the factor to be used in the calculation of overtrade amounts.}

\item{nr_pc}{An integer, the number of eigen portfolios to extract from the returns data.}
}
\value{
A numeric matrix of predicted returns where each row corresponds to a day in the given `start:end` range and each column corresponds to a stock.
}
\description{
This function performs cross-sectional regression on a range of days to predict returns, considering overtraded volume. 
For each day in the range, it standardises the volume, calculates overtrade amounts, adjusts the returns by the 
exponential of the overtrade amounts, and fits a linear model. The function utilizes parallel computing to increase speed.
}
\examples{
#Example data
volume <- matrix(rnorm(25), 5, 5)
returns <- matrix(rnorm(25), 5, 5)
start <- 2
end <- 5
h <- 4
nr_pc_v <- 2
alpha <- 0.05
nr_pc <- 2
#Use the function
crossSectionalRegressionOvertraded(start, end, volume, returns, h, nr_pc_v, alpha, nr_pc)

}
