% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross-sectional-regression-vw.R
\name{crossSectionalRegressionVW}
\alias{crossSectionalRegressionVW}
\title{Cross-Sectional Regression with Volume Weighting}
\usage{
crossSectionalRegressionVW(returns, volume, start, end, h, nr_pc, d, divide)
}
\arguments{
\item{returns}{A numeric matrix, the returns data.}

\item{volume}{A numeric matrix, the volume data.}

\item{start}{An integer, the start of the days range for which to perform regression.}

\item{end}{An integer, the end of the days range for which to perform regression.}

\item{h}{An integer, the number of recent historical days to be used.}

\item{nr_pc}{An integer, the number of eigen portfolios to extract.}

\item{d}{An integer, the width for rolling mean calculation on volume.}

\item{divide}{A logical value, if TRUE, the function divides the product of returns and rolling mean volume by the volume. If FALSE, it divides the product of returns and volume by the rolling mean volume.}
}
\value{
A numeric matrix of predicted returns where each row corresponds to a day in the given `start:end` range and each column corresponds to a stock.
}
\description{
This function performs cross-sectional regression on volume-weighted returns on a range of days,
using recent historical data. An eigenportfolio is first extracted from the historical data.
Then, a linear model is fit between the returns and the eigenportfolio for each day in the range.
The function returns the negative residuals from these models as predicted returns. 
This is a parallelized function that takes advantage of multiple cores for speed.
}
\examples{
#Example data
returns <- matrix(rnorm(25), 5, 5)
volume <- matrix(rnorm(25), 5, 5)
start <- 2
end <- 5
h <- 4
nr_pc <- 2
d <- 2
divide <- TRUE
#Use the function
crossSectionalRegressionVW(returns, volume, start, end, h, nr_pc, d, divide)
}
