% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross-sectional-regression-kcca.R
\name{crossSectionalRegressionWithKCCA}
\alias{crossSectionalRegressionWithKCCA}
\title{Cross-Sectional Regression with Kernel Canonical Correlation Analysis (KCCA)}
\usage{
crossSectionalRegressionWithKCCA(returns, volume, t, h, d, hv, nr_c_r, nr_c_v)
}
\arguments{
\item{returns}{A numeric matrix, the returns data.}

\item{volume}{A numeric matrix, the volume data.}

\item{t}{An integer, the day for which to perform regression.}

\item{h}{An integer, number of recent historical days to be used for returns data.}

\item{d}{An integer, the width for rolling mean calculation on volume.}

\item{hv}{An integer, number of recent historical days to be used for volume data.}

\item{nr_c_r}{An integer, the number components of returns to be used in KCCA.}

\item{nr_c_v}{An integer, the number components of volume to be used in KCCA.}
}
\value{
A numeric matrix of predicted returns where each row corresponds to a day in the given `t` range and each column corresponds to a stock.
}
\description{
This function performs cross-sectional regression on a range of days to predict returns, considering the effect of volume. 
For each day in the range, it standardizes the volume, applies KCCA to find correlation between the returns and volume data,
and fits a linear model. The residuals from the model are returned as predicted returns. The function is parallelized for increased speed.
}
\examples{
#Example data
returns <- matrix(rnorm(20), 4, 5)
volume <- matrix(rnorm(20), 4, 5)
t <- 5
h <- 2
d <- 2
hv <- 2
nr_c_r <- 2
nr_c_v <- 2
#Use the function
crossSectionalRegressionWithKCCA(returns, volume, t, h, d, hv, nr_c_r, nr_c_v)

}
