% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross-temporal-regression.R
\name{crossTemporalRegression}
\alias{crossTemporalRegression}
\title{Cross-Temporal Regression}
\usage{
crossTemporalRegression(returns, start, end, nr_pc, h, l, b_sensitivity)
}
\arguments{
\item{returns}{A numeric matrix, the returns data.}

\item{start}{An integer, the start of the date range for which to calculate S-Scores.}

\item{end}{An integer, the end of the date range for which to calculate S-Scores.}

\item{nr_pc}{An integer, the number of eigen portfolios to extract.}

\item{h}{An integer, the number of recent historical days to be used.}

\item{l}{An integer, the number of preceding days used for the model.}

\item{b_sensitivity}{A numeric value, sensitivity for checking if model b coefficient is less than 1 minus this value.}
}
\value{
A matrix of S-Scores where each row corresponds to a stock and each column corresponds 
to a date in the given `start:end` range.
}
\description{
This function is a parallelized process that calculates S-Scores for each day in a given date 
range using only the historical data upto that date. The S-Score is a measure of mean reversion
tendency for each stock, and is calculated by `calculateSScore` function. Multiple packages 
are used to conduct parallel processing for efficiency.
}
\examples{
#Example data
returns <- matrix(rnorm(25), 5, 5)
nr_pc <- 2
h <- 4
l <- 1
b_sensitivity <- 0.1
start <- 2
end <- 5
#Use the function
crossTemporalRegression(returns, start, end, nr_pc, h, l, b_sensitivity)
}
