% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross-temporal-regression-cca.R
\name{crossTemporalRegressionCCA}
\alias{crossTemporalRegressionCCA}
\title{Cross-Temporal Regression With Canonical Correlation Analysis (CCA)}
\usage{
crossTemporalRegressionCCA(
  returns,
  volume,
  start,
  end,
  h,
  hv,
  l,
  nr_c_r,
  nr_c_v,
  d,
  b_sensitivity
)
}
\arguments{
\item{returns}{A numeric matrix, the returns data.}

\item{volume}{A numeric matrix, the volume data.}

\item{start}{An integer, the start of the date range for which to calculate S-Scores.}

\item{end}{An integer, the end of the date range for which to calculate S-Scores.}

\item{h}{An integer, the number of recent historical days to be used for returns.}

\item{hv}{An integer, the number of recent historical days to be used for standardised volume.}

\item{l}{An integer, the number of preceding days used for the model.}

\item{nr_c_r}{An integer, the number of components used for returns in CCA.}

\item{nr_c_v}{An integer, the number of components used for standardised volume in CCA.}

\item{d}{An integer, the width for calculating the rolling mean of the volume.}

\item{b_sensitivity}{A numeric value, sensitivity for checking if model b coefficient is less than 1 minus this value.}
}
\value{
A matrix of S-Scores where each row corresponds to a stock and each column corresponds to a date in the given `start:end` range.
}
\description{
This function performs cross-temporal regression on a range of days using CCA analysis.
It is a parallelized process that calculates S-Scores for each day based on volume-standardised 
historical returns. CCA is applied to determine the correlation between returns and standardised volume.
}
\examples{
#Example data
returns <- matrix(rnorm(25), 5, 5)
volume <- matrix(rnorm(25), 5, 5)
start <- 2
end <- 5
hurns <- 4
hv <- 4
l <- 1
nr_c_r <- 2
nr_c_v <- 2
d <- 2
b_sensitivity <- 0.1
#Use the function
crossTemporalRegressionCCA(returns, volume, start, end, h, hv, l, nr_c_r, nr_c_v, d, b_sensitivity)
}
