% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross-temporal-regression-overtraded.R
\name{crossTemporalRegressionOvertrade}
\alias{crossTemporalRegressionOvertrade}
\title{Cross-Temporal Regression With Overtrade}
\usage{
crossTemporalRegressionOvertrade(
  volume,
  returns,
  start,
  end,
  h,
  nr_pc.V,
  nr_pc,
  alpha,
  l,
  b_sensitivity
)
}
\arguments{
\item{volume}{A numeric matrix, the volume data.}

\item{returns}{A numeric matrix, the returns data.}

\item{start}{An integer, the start of the date range for which to calculate S-Scores.}

\item{end}{An integer, the end of the date range for which to calculate S-Scores.}

\item{h}{An integer, the number of recent historical days to be used.}

\item{nr_pc.V}{An integer, the number of eigen portfolios to extract based on volume.}

\item{nr_pc}{An integer, the number of eigen portfolios to extract based on returns.}

\item{alpha}{A numeric, the weight on overtrading term.}

\item{l}{An integer, the number of preceding days used for the model.}

\item{b_sensitivity}{A numeric value, sensitivity for checking if model b coefficient is less than 1 minus this value.}
}
\value{
A matrix of S-Scores where each row corresponds to a stock and each column corresponds to a date in the given `start:end` range.
}
\description{
This function performs cross-temporal regression on a range of days with consideration for overtrading.
It is a parallelized process that calculates S-Scores for each day based on volume-weighted historical returns.
Overtrading is accounted for by adjusting the standardised volume with the `singleCrossTemporalRegressionOvertrade` function.
}
\examples{
#Example data
returns <- matrix(rnorm(25), 5, 5)
volume <- matrix(rnorm(25), 5, 5)
start <- 2
end <- 5
t <- 6
h <- 4
preceding <- 1
nr_pc.V <- 2
nr_pc <- 2
alpha <- 0.1
l <- 2
b_sensitivity <- 0.1
#Use the function
crossTemporalRegressionOvertrade(volume, returns, start, end, t, h, nr_pc.V, nr_pc, alpha, preceding, b_sensitivity)

}
