% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross-temporal-regression-kcca.R
\name{crossTemporalRegressionWithKCCA}
\alias{crossTemporalRegressionWithKCCA}
\title{Cross-Temporal Regression With Kernel Canonical Correlation Analysis (KCCA)}
\usage{
crossTemporalRegressionWithKCCA(
  returns,
  volume,
  start,
  end,
  h,
  hv,
  l,
  nr_c_r,
  nr_c_v,
  d,
  b_sensitivity
)
}
\arguments{
\item{returns}{A numeric matrix, the returns data.}

\item{volume}{A numeric matrix, the volume data.}

\item{start}{An integer, the start of the date range for which to calculate S-Scores.}

\item{end}{An integer, the end of the date range for which to calculate S-Scores.}

\item{h}{An integer, the number of recent historical days to be used for returns.}

\item{hv}{An integer, the number of recent historical days to be used for volume.}

\item{l}{An integer, the number of preceding days used for the model.}

\item{nr_c_r}{An integer, the number of components used for returns in KCCA.}

\item{nr_c_v}{An integer, the number of components used for volume in KCCA.}

\item{d}{An integer, the width for rolling mean calculation on volume.}

\item{b_sensitivity}{A numeric value, sensitivity for checking if model b coefficient is less than 1 minus this value.}
}
\value{
A matrix of S-Scores where each row corresponds to a stock and each column corresponds to a date in the given `start:end` range.
}
\description{
This function performs cross-temporal regression on a range of days using KCCA analysis.
It is a parallelized process that calculates S-Scores for each day and each stock based on volume-weighted historical returns.
KCCA is applied to determine the correlation between returns and volume.
}
\examples{
#Example data
returns <- matrix(rnorm(25), 5, 5)
volume <- matrix(rnorm(25), 5, 5)
start <- 2
end <- 5
h <- 4
hb <- 4
l <- 1
nr_c_r <- 2
nr_c_v <- 2
d <- 2
b_sensitivity <- 0.1
#Use the function
crossTemporalRegressionWithKCCA(returns, volume, start, end, h, hv, l, nr_c_r, nr_c_v, d, b_sensitivity)

}
