% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross-temporal-regression-vw.R
\name{crossTemporalRegressionWithVW}
\alias{crossTemporalRegressionWithVW}
\title{Add together two numbers}
\usage{
crossTemporalRegressionWithVW(
  returns,
  volume,
  start,
  end,
  nr_pc,
  h,
  l,
  b_sensitivity,
  d,
  divide
)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
A numeric vector.
}
\description{
Add together two numbers
}
\examples{
add(1, 1)
add(10, 1)
}
