% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-s-score.R
\name{estimateCoefficeients}
\alias{estimateCoefficeients}
\title{Estimate Key Coefficients For All Models}
\usage{
estimateCoefficeients(models, b_sensitivity)
}
\arguments{
\item{models}{A list of linear models for each stock, as obtained from the `decompose` function.}

\item{b_sensitivity}{A numeric value, the sensitivity for b parameter to declare model as mean reverting.}
}
\value{
A data frame where each row corresponds to the estimated key coefficients of each model: 
'k' for speed of mean reversion, 'm' for the long term mean, 'sigma_squared' for variance, 
'sigma_eq_squared' for equilibrium variance and 'is_mean_reverting' for the model's mean-reverting status.
}
\description{
This function applies the `estimateSingleCoefficientsNumber` function to all models
in the list. After key coefficients are calculated for each model, results are 
combined in a single data frame.
}
\examples{
#Example data
ret <- matrix(rnorm(25), 5, 5)
num_pc <- 2
historical <- 4
preceding <- 1
sensitivity <- 0.1
#Use the function
models <- decompose(ret, num_pc, historical, preceding)
estimateCoefficients(models, sensitivity)

}
