% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-s-score.R
\name{estimateSingleCoefficeientsNumber}
\alias{estimateSingleCoefficeientsNumber}
\title{Estimates Mean Reverting Coefficients}
\usage{
estimateSingleCoefficeientsNumber(models, i, b_sensitivity)
}
\arguments{
\item{models}{A list of linear models for each stock, as obtained from the `decompose` function.}

\item{i}{An integer, the index of the specific model required from the list.}

\item{b_sensitivity}{A numeric value, sensitivity for checking if model_b < 1 - b_sensitivity.}
}
\value{
A numeric vector of length 5 containing key model parameters.
}
\description{
This function estimates the key coefficients in a mean reverting model. 
A linear model is created first using cumulative sum of residuals as variables. 
Then, values for speed of mean reversion (k), long term mean (m), variance (sigma squared),
equilibrium variance (sigma_eq_squared) and mean-reverting indicator are calculated.
}
\examples{
#Example data
ret <- matrix(rnorm(25), 5, 5)
num_pc <- 2
historical <- 4
preceding <- 1
sensitivity <- 0.1
#Use the function
models <- decompose(ret, num_pc, historical, preceding)
estimateCoefficeientsNumberI(models, 1, sensitivity)

}
