% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-eigen.R
\name{extractEigenPortfolios}
\alias{extractEigenPortfolios}
\title{Extract Eigen Portfolios}
\usage{
extractEigenPortfolios(returns, nr_pc)
}
\arguments{
\item{returns}{A numeric matrix, the returns data.}

\item{nr_pc}{An integer, the number of eigen portfolios to extract.}
}
\value{
A list containing:
* `rho`: The correlation matrix for the returns.
* `portfolio`: The selected eigen portfolios.
* `return`: The returns of the selected eigen portfolios.
* `prop_explained`: The proportion of variance explained by the selected eigen portfolios.
}
\description{
This function extracts a specified number of eigen portfolios from a matrix of returns. 
It uses the `constructEigenPortfolios` function to first create all eigen portfolios, 
then selects the first `nr_pc` portfolios based on the eigenvalues.
}
\examples{
#Example data
returns <- matrix(rnorm(25), 5, 5)
x <- 2
#Use the function
extractEigenPortfolios(returns,  x)
}
