% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross-temporal-regression-vw.R
\name{mappedCrossTemporalRegressionWithVW}
\alias{mappedCrossTemporalRegressionWithVW}
\title{Mapped Cross-Temporal Regression With Volume Weighting}
\usage{
mappedCrossTemporalRegressionWithVW(
  returns,
  volume,
  start,
  end,
  nr_pc,
  h,
  l,
  b_sensitivity,
  d,
  divide,
  maps
)
}
\arguments{
\item{returns}{A numeric matrix, the returns data.}

\item{volume}{A numeric vector, the volume data.}

\item{start}{An integer, the start of the date range for which to calculate S-Scores.}

\item{end}{An integer, the end of the date range for which to calculate S-Scores.}

\item{nr_pc}{An integer, the number of eigen portfolios to extract.}

\item{h}{An integer, the number of recent historical days to be used.}

\item{l}{An integer, the number of preceding days used for the model.}

\item{b_sensitivity}{A numeric value, sensitivity for checking if model b coefficient is less than 1 minus this value.}

\item{d}{An integer, the number of days used for volume weighting.}

\item{divide}{A logical value, indicating whether to divide returns by volume score.}

\item{maps}{A list of functions, each function is used to map the volume data.}
}
\value{
A list of S-Score matrices (from `crossTemporalRegressionWithVW`), one for each mapping function.
}
\description{
This function maps the volume data using given maps before performing cross-temporal 
regression with volume weighting. It calculates S-Scores for each day using volume weighted 
historical data for each map.
}
\examples{
#Example data
returns <- matrix(rnorm(25), 5, 5)
volume <- rnorm(5)
nr_pc <- 2
h <- 4
l <- 1
b_sensitivity <- 0.1
start <- 2
end <- 5
d <- 2
divide <- TRUE
maps <- list(sqrt, log)
#Use the function
mappedCrossTemporalRegressionWithVW(returns, volume, start, end, nr_pc, h, l, b_sensitivity, d, divide, maps)
}
