% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross-sectional-regression.R
\name{singleCrossSectionalRegression}
\alias{singleCrossSectionalRegression}
\title{Single Cross-Sectional Regression}
\usage{
singleCrossSectionalRegression(returns, t, h, nr_pc)
}
\arguments{
\item{returns}{A numeric matrix, the returns data.}

\item{t}{An integer, the day for which to perform regression.}

\item{h}{An integer, the number of recent historical days to be used.}

\item{nr_pc}{An integer, the number of eigen portfolios to extract.}
}
\value{
A numeric vector of predictions for returns on day `t`, derived from the cross-sectional regression model.
}
\description{
This function performs cross-sectional regression on returns on a given day,
using recent historical data. An eigenportfolio is first extracted from the historical data.
Then, a linear model is fit between the returns on the day and the eigenportfolio.
The function returns the negative residuals from this model which denote the predicted returns.
}
\examples{
#Example data
returns <- matrix(rnorm(20), 4, 5)
t <- 5
h <- 4
nr_pc <- 2
#Use the function
singleCrossSectionalRegression(returns, t, h, nr_pc)
}
