% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross-sectional-regression-clustering.R
\name{singleCrossSectionalRegressionClustering}
\alias{singleCrossSectionalRegressionClustering}
\title{Single Cross-Sectional Regression with Clustering}
\usage{
singleCrossSectionalRegressionClustering(
  returns,
  volume,
  t,
  h,
  nr_pc,
  k,
  min_size,
  alpha
)
}
\arguments{
\item{returns}{A numeric matrix, the returns data.}

\item{volume}{A numeric matrix, the volume data.}

\item{t}{An integer, the day for which to perform regression.}

\item{h}{An integer, the number of recent historical days to be used.}

\item{nr_pc}{An integer, the number of principal components to extract.}

\item{k}{An integer, the desired number of clusters.}

\item{min_size}{An integer, the minimum desired size of the clusters.}

\item{alpha}{A numeric, the weight given to the correlation matrix in the combined similarity matrix used for clustering.}
}
\value{
A list with two elements:
        - A numeric vector of normalized predictions for returns on day `t`.
        - The final number of clusters.
}
\description{
This function performs cross-sectional regression on returns on a given day,
using recent historical data. Data points are clustered based on correlation and trading volume.
For each cluster, a linear model is fit between the returns of the previous day and the eigenportfolios. 
The negative residuals from these models are returned as predicted returns.
}
\examples{
#Example data
returns <- matrix(rnorm(25), 5, 5)
volume <- matrix(rnorm(25), 5, 5)
t <- 3
h <- 2
nr_pc <- 2
k <- 2
min_size <- 2
alpha <- 0.5
#Use the function
singleCrossSectionalRegressionClustering(returns, volume, t, h, nr_pc, k, min_size, alpha)

}
