% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross-sectional-regression-kpca.R
\name{singleCrossSectionalRegressionKPCA}
\alias{singleCrossSectionalRegressionKPCA}
\title{Single Cross-Sectional Regression with Kernel Principal Component Analysis (KPCA)}
\usage{
singleCrossSectionalRegressionKPCA(returns, t, h, nr_pc, kernel, kpar)
}
\arguments{
\item{returns}{A numeric matrix, the returns data.}

\item{t}{An integer, the day for which to perform regression.}

\item{h}{An integer, the number of recent historical days to be used.}

\item{nr_pc}{An integer, the number of principal components to keep in KPCA.}

\item{kernel}{A character string, the kernel to be used in KPCA.}

\item{kpar}{A list, the list of hyperparameters for the kernel.}
}
\value{
A numeric vector of predictions for returns on day `t`, derived from the regression model using KPCA.
}
\description{
This function performs cross-sectional regression on scaled returns on a given day,
using recent historical data. KPCA is applied to the scaled returns. Then, a linear 
model is fit between the last day's returns and the KPCA rotated returns. 
The function returns the negative residuals from this model which denote the predicted returns.
}
\examples{
#Example data
returns <- matrix(rnorm(20), 4, 5)
t <- 5
h <- 4
nr_pc <- 2
kernel <- "rbfdot"
kpar <- list(sigma = 0.1)
#Use the function
singleCrossSectionalRegressionKPCA(returns, t, h, nr_pc, kernel, kpar
}
