% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross-sectional-regression-overtrade.R
\name{singleCrossSectionalRegressionOvertraded}
\alias{singleCrossSectionalRegressionOvertraded}
\title{Single Cross-Sectional Regression for Overtraded Volume}
\usage{
singleCrossSectionalRegressionOvertraded(
  volume,
  returns,
  t,
  h,
  nr_pc_v,
  nr_pc,
  alpha
)
}
\arguments{
\item{volume}{A numeric matrix, the volume data.}

\item{returns}{A numeric matrix, the returns data.}

\item{t}{An integer, the day for which to perform regression.}

\item{h}{An integer, the number of recent historical days to be used.}

\item{nr_pc_v}{An integer, the number of eigen portfolios to extract from the volume data.}

\item{nr_pc}{An integer, the number of eigen portfolios to extract from the returns data.}

\item{alpha}{A numeric, the factor to be used in the calculation of overtrade amounts.}
}
\value{
A numeric vector of residuals from the cross-sectional regression model for the day `t`.
}
\description{
This function performs cross-sectional regression to predict returns, taking into account overtraded volume. 
It first standardises the volume and extracts an eigenportfolio. Then, it calculates the overtrade amounts 
by fitting a linear model between the standardised volume and the eigenportfolio. Following this, it adjusts 
the returns by the exponential of the overtrade amounts, and fits another linear model between the adjusted 
returns and a new eigenportfolio. The function returns the negative residuals from this model.
}
\examples{
#Example data
volume <- matrix(rnorm(25), 5, 5)
returns <- matrix(rnorm(25), 5, 5)
t <- 5
h <- 4
nr_pc_v <- 2
nr_pc_r <- 2
alpha <- 0.05
#Use the function
singleCrossSectionalRegressionOvertraded(volume, returns, t, h, nr_pc_v, nr_pc_v, alpha)

}
