% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross-sectional-regression-kcca.R
\name{singleCrossSectionalRegressionWithKCCA}
\alias{singleCrossSectionalRegressionWithKCCA}
\title{Single Cross-Sectional Regression with Kernel Canonical Correlation Analysis (KCCA)}
\usage{
singleCrossSectionalRegressionWithKCCA(
  returns,
  volume,
  t,
  h,
  hv,
  nr_c_r,
  nr_c_v
)
}
\arguments{
\item{returns}{A numeric matrix, the returns data.}

\item{volume}{A numeric matrix, the volume data.}

\item{t}{An integer, the day for which to perform regression.}

\item{h}{An integer, number of recent historical days to be used for returns data.}

\item{hv}{An integer, number of recent historical days to be used for volume data.}

\item{nr_c_r}{An integer, the number components of returns to be used in KCCA.}

\item{nr_c_v}{An integer, the number components of volume to be used in KCCA.}
}
\value{
A numeric vector of predictions for returns on day `t`, derived from the regression model using KCCA.
}
\description{
This function performs cross-sectional regression on scaled returns on a given day,
using recent historical data. It uses KCCA to find correlation between the returns 
and volume data. Then, a linear model is fit between the returns of the previous day
and the obtained KCCA coefficients of returns and volume. The function returns the 
negative residuals from this model as predicted returns.
}
\examples{
#Example data
returns <- matrix(rnorm(20), 4, 5)
volume <- matrix(rnorm(20), 4, 5)
t <- 5
h <- 2
hv <- 2
nr_c_r <- 2
nr_c_v <- 2
#Use the function
singleCrossSectionalRegressionWithKCCA(returns, volume, t, h, hv, nr_c_r, nr_c_v)

}
