% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross-temporal-regression-cca.R
\name{singleCrossTemporalRegressionCCA}
\alias{singleCrossTemporalRegressionCCA}
\title{Single Cross-Temporal Regression With Canonical Correlation Analysis (CCA)}
\usage{
singleCrossTemporalRegressionCCA(
  returns,
  standardised_volume,
  t,
  h,
  hv,
  l,
  nr_c_r,
  nr_c_v,
  b_sensitivity
)
}
\arguments{
\item{returns}{A numeric matrix, the returns data.}

\item{standardised_volume}{A numeric matrix, the standardised_volume data.}

\item{t}{An integer, the current day for which to calculate S-Scores.}

\item{h}{An integer, the number of recent historical days to be used for returns.}

\item{hv}{An integer, the number of recent historical days to be used for standardised volume.}

\item{l}{An integer, the number of preceding days used for the model.}

\item{nr_c_r}{An integer, the number of components used for returns in CCA.}

\item{nr_c_v}{An integer, the number of components used for standardised volume in CCA.}

\item{b_sensitivity}{A numeric value, sensitivity for checking if model b coefficient is less than 1 minus this value.}
}
\value{
A list containing:
* `s`: The computed S-Score for each stock.
* `is_mean_reverting`: A logical vector indicating whether each stock is mean reverting.
}
\description{
This function applies Canonical Correlation Analysis (CCA) to determine 
the correlation between returns and standardised volume. Then, it performs cross-temporal regression 
on the given day and returns an S-Score, which measures the mean reversion tendency for each stock.
}
\examples{
#Example data
returns <- matrix(rnorm(25), 5, 5)
volume <- matrix(rnorm(25), 5, 5)
standardised_volume <- vol / rowMeans(vol)
t <- 6
h <- 4
hv <- 4
l <- 1
nr_c_r <- 2
nr_c_v <- 2
b_sensitivity <- 0.1
#Use the function
singleCrossTemporalRegressionCCA(returns, standardised_volume, t, h, hv, l, nr_c_r, nr_c_v, b_sensitivity)
}
