% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross-temporal-regression-overtraded.R
\name{singleCrossTemporalRegressionOvertrade}
\alias{singleCrossTemporalRegressionOvertrade}
\title{Single Cross-Temporal Regression Overtrade}
\usage{
singleCrossTemporalRegressionOvertrade(
  volume,
  returns,
  t,
  h,
  nr_pc.V,
  nr_pc,
  alpha,
  l,
  b_sensitivity
)
}
\arguments{
\item{volume}{A numeric matrix, the volume data.}

\item{returns}{A numeric matrix, the returns data.}

\item{t}{An integer, the current day for which to calculate S-Scores.}

\item{h}{An integer, the number of recent historical days to be used.}

\item{nr_pc.V}{An integer, the number of eigen portfolios to construct based on volume.}

\item{nr_pc}{An integer, the number of eigen portfolios to construct based on returns.}

\item{alpha}{A numeric, the weight on overtrading term.}

\item{l}{An integer, the number of preceding days used for the model.}

\item{b_sensitivity}{A numeric value, sensitivity for checking if model b coefficient is less than 1 minus this value.}
}
\value{
A list containing:
* `s`: The computed S-Score for each stock.
* `is_mean_reverting`: A logical vector indicating whether each stock is mean reverting.
}
\description{
This function performs cross-temporal regression with consideration for overtrading.
It standardizes volume using recent historical data, and then constructs the eigen portfolios 
based on this standardized volume. The function then decomposes the volume weighted returns into 
a mean-reverting model for each stock and calculates the S-Score.
}
\examples{
#Example data
returns <- matrix(rnorm(25), 5, 5)
volume <- matrix(rnorm(25), 5, 5)
t <- 6
h <- 4
preceding <- 1
nr_pc.V <- 2
nr_pc <- 2
alpha <- 0.1
l <- 2
b_sensitivity <- 0.1
#Use the function
singleCrossTemporalRegressionOvertrade(volume, returns, t, h, nr_pc.V, nr_pc, alpha, preceding, b_sensitivity)
}
