% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross-temporal-regression-kcca.R
\name{singleCrossTemporalRegressionWithKCCA}
\alias{singleCrossTemporalRegressionWithKCCA}
\title{Single Cross-Temporal Regression With Kernel Canonical Correlation Analysis (KCCA)}
\usage{
singleCrossTemporalRegressionWithKCCA(
  returns,
  volume,
  t,
  h,
  hv,
  l,
  nr_c_r,
  nr_c_v,
  b_sensitivity
)
}
\arguments{
\item{returns}{A numeric matrix, the returns data.}

\item{volume}{A numeric matrix, the volume data.}

\item{t}{An integer, the current day for which to calculate S-Scores.}

\item{h}{An integer, the number of recent historical days to be used for returns.}

\item{hv}{An integer, the number of recent historical days to be used for volume.}

\item{l}{An integer, the number of preceding days used for the model.}

\item{nr_c_r}{An integer, the number of components used for returns in KCCA.}

\item{nr_c_v}{An integer, the number of components used for volume in KCCA.}

\item{b_sensitivity}{A numeric value, sensitivity for checking if model b coefficient is less than 1 minus this value.}
}
\value{
A list containing:
* `s`: The computed S-Score for each stock.
* `is_mean_reverting`: A logical vector indicating whether each stock is mean reverting.
}
\description{
This function applies Kernel Canonical Correlation Analysis (KCCA) to determine 
the correlation between returns and volume. Then, it performs cross-temporal regression 
on the given day and returns an S-Score, which measures the mean reversion tendency for each stock.
}
\examples{
#Example data
returns <- matrix(rnorm(25), 5, 5)
volume <- matrix(rnorm(25), 5, 5)
t_val <- 6
hurns <- 4
historical_volume <- 4
preceding <- 1
kcca_components_returns <- 2
kcca_components_volume <- 2
sensitivity <- 0.1
#Use the function
singleCrossTemporalRegressionWithKCCA(returns, volume, t_val, h, historical_volume, preceding, kcca_components_ret, kcca_components_vol, sensitivity)
}
