% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross-temporal-regression-vw.R
\name{crossTemporalRegressionWithVW}
\alias{crossTemporalRegressionWithVW}
\title{Cross-Temporal Regression With Volume Weighting}
\usage{
crossTemporalRegressionWithVW(
  returns,
  volume,
  start,
  end,
  nr_pc,
  h,
  l,
  b_sensitivity,
  d,
  divide
)
}
\arguments{
\item{returns}{A numeric matrix, the returns data.}

\item{volume}{A numeric vector, the volume data.}

\item{start}{An integer, the start of the date range for which to calculate S-Scores.}

\item{end}{An integer, the end of the date range for which to calculate S-Scores.}

\item{nr_pc}{An integer, the number of eigen portfolios to extract.}

\item{h}{An integer, the number of recent historical days to be used.}

\item{l}{An integer, the number of preceding days used for the model.}

\item{b_sensitivity}{A numeric value, sensitivity for checking if model b coefficient is less than 1 minus this value.}

\item{d}{An integer, the number of days used for volume weighting.}

\item{divide}{A logical value, indicating whether to divide returns by volume score.}
}
\value{
A matrix of S-Scores where each row corresponds to a stock and each column corresponds 
to a date in the given `start:end` range.
}
\description{
This function is a parallelized process that calculates S-Scores for each day in a given date 
range using only the volume weighted historical data up to that date. The S-Score is a measure of 
mean reversion tendency for each stock, and is calculated by the `calculateSScore` function.
The function uses multiple packages to conduct parallel processing for efficiency.
}
\examples{
#Example data
returns <- matrix(rnorm(25), 5, 5)
volume <- rnorm(5)
nr_pc <- 2
h <- 4
l <- 1
b_sensitivity <- 0.1
start <- 2
end <- 5
d <- 2
divide <- TRUE
#Use the function
crossTemporalRegressionWithVW(returns, volume, start, end, nr_pc, h, l, b_sensitivity, d, divide)
}
